Move rpcap-protocol.h to the top-level directory.
That means that everything in rpcap is part of rpcapd.

Fix up the list of files included in the tarball while we're at it.
