+++ pcap-new.c
@@ -37,7 +37,7 @@
 
 #include "pcap-int.h"	// for the details of the pcap_t structure
 #include "sockutils.h"
-#include "rpcap/rpcap-protocol.h"
+#include "rpcap-protocol.h"
 #include "pcap-rpcap-int.h"
 #include <errno.h>		// for the errno variable
 #include <stdlib.h>		// for malloc(), free(), ...
